Files for PHP Backend
These files connect to the database, and perform the operation which their name suggests. Ensure that there is a file connect.php under the config directory of the project; which can help you to connect to the database via PHP. Also ensure to use the .htaccess file which I uploaded , which translates update -> update.php; upload -> upload.php and so on.